+++ wrappers/python/src/checkrd/integrations/__init__.py
@@ -9,7 +9,6 @@ from checkrd.integrations._base import HttpxClientInstrumentor, Instrumentor
 from checkrd.integrations._cohere import CohereInstrumentor
 from checkrd.integrations._google_genai import GoogleGenAIInstrumentor
 from checkrd.integrations._groq import GroqInstrumentor
-from checkrd.integrations._mistralai import MistralInstrumentor
 from checkrd.integrations._openai import OpenAIInstrumentor
 from checkrd.integrations._together import TogetherInstrumentor
 
@@ -19,7 +18,6 @@ __all__ = [
     "OpenAIInstrumentor",
     "AnthropicInstrumentor",
     "CohereInstrumentor",
-    "MistralInstrumentor",
     "GroqInstrumentor",
     "TogetherInstrumentor",
     "GoogleGenAIInstrumentor",